requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
gevent==23.9.1
greenlet==3.0.0rc3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
Naked==0.1.32
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycryptodome==3.19.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
